Move `serde_json` benchmarks into their own file.
